Extended personalized build script

+++ x/build.ps1
@@ -20,40 +20,52 @@ function NormalizePath {
 $thisFolder = Split-Path -Parent $MyInvocation.MyCommand.Definition | NormalizePath
 $rootFolder = Split-Path -Parent $thisFolder | NormalizePath
 
+$fresh = $false
+$cuda = $true
+
 # Optimized for AMD Ryzen 9 3900X (Zen 2, 12 cores, 24 threads)
 if ($IsWindows) {
   $cFlags = '/arch:AVX2 /favor:AMD64 /fp:fast /MP /O2 /Oi /Ot'
-}
-else {
+} else {
   $cFlags = '-march=znver2 -mtune=znver2 -O3'
 }
 $flags = @{
   'CMAKE_BUILD_TYPE' = 'Release'
   'CMAKE_C_COMPILER' = 'clang'
   'CMAKE_C_FLAGS' = $cFlags
-  'CMAKE_CUDA_ARCHITECTURES' = '89' # Optimized for Nvidia GeForce RTX 4070
-  'CMAKE_CUDA_COMPILER_TOOLKIT_ROOT' = $env:CUDA_PATH
-  'CMAKE_CUDA_FLAGS' = '-t6'
   'CMAKE_CXX_COMPILER' = 'clang++'
   'CMAKE_CXX_FLAGS' = $cFlags
   'CMAKE_POSITION_INDEPENDENT_CODE' = '1'
-  'GGML_CUDA_F16' = '1'
-  'GGML_CUDA_FA_ALL_QUANTS' = '1'
-  'GGML_CUDA_NO_PEER_COPY' = '1'
-  'GGML_CUDA' = '1'
   'GGML_NATIVE' = '1'
-  'GGML_SYCL_TARGET' = 'NVIDIA'
 }
+if ($cuda) {
+  $flags += @{
+    'CMAKE_CUDA_ARCHITECTURES' = '89'
+    'CMAKE_CUDA_COMPILER_TOOLKIT_ROOT' = $env:CUDA_PATH
+    'CMAKE_CUDA_FLAGS' = '-t6'
+    'GGML_CUDA_F16' = '1'
+    'GGML_CUDA_FA_ALL_QUANTS' = '1'
+    'GGML_CUDA_NO_PEER_COPY' = '1'
+    'GGML_CUDA' = '1'
+    'GGML_SYCL_TARGET' = 'NVIDIA'
+  }
+} else {
+  $flags += @{
+    'GGML_CUDA' = '0'
+  }
+}
+$env:__OPTIMIZE__ = '1'
 
 $buildFolder = "$rootFolder/build" | NormalizePath
 $buildFolderExists = Test-Path $buildFolder
 if ($buildFolderExists) {
-  Remove-Item -Recurse -Force $buildFolder
-}
-else {
+  if ($fresh) {
+    Remove-Item -Recurse -Force $buildFolder
+    New-Item -ItemType Directory -Path $buildFolder | Out-Null
+  }
+} else {
   New-Item -ItemType Directory -Path $buildFolder | Out-Null
 }
-
 function Configure {
   Set-Location $buildFolder
   $cmakeCommand = 'cmake'
@@ -73,11 +85,10 @@ function Configure {
   Write-Output "Generated CMake Command: $cmakeCommand"
   Invoke-Expression $cmakeCommand
 }
-
 function Compile {
   Set-Location $buildFolder
-  $env:CMAKE_BUILD_PARALLEL_LEVEL = 8
-  cmake --build . --config Release
+  $env:CMAKE_BUILD_PARALLEL_LEVEL = 12
+  cmake --build . --config Release --target llama-cli llama-imatrix llama-perplexity llama-quantize llama-server
 }
 
 Configure
